+++ crates/voicevox_core_c_api/tests/e2e/testcases/user_dict_load.rs
@@ -134,7 +134,13 @@ impl assert_cdylib::TestCase for TestCase {
         };
 
         // SAFETY: `voicevox_synthesizer_load_voice_model` has no safety requirements.
-        assert_ok(unsafe { lib.voicevox_synthesizer_load_voice_model(synthesizer, model) });
+        assert_ok(unsafe {
+            lib.voicevox_synthesizer_load_voice_model(
+                synthesizer,
+                model,
+                lib.voicevox_make_default_load_voice_model_options(),
+            )
+        });
 
         let mut audio_query_without_dict = std::ptr::null_mut();
         assert_ok(unsafe {