+++ CHANGELOG.md
@@ -55,7 +55,7 @@
 
 ### Added
 
-- \[Rust\] `Synthesizer::load_voice_model`にオプション`on_existing`が追加されます ([#1331])。
+- `Synthesizer::load_voice_model`にオプション`on_existing`が追加されます ([#1331], [#1337])。
 - \[ダウンローダー\] `--os`オプションで`android`と`ios`を指定できるようになります。ただしiOSの`c-api`をダウンロードすることはできません ([#1313])。
 - \[ダウンローダー\] 環境変数`VV_DOWNLOADER_C_API_ALLOW_DRAFT`を設定することで、`c-api`のdraft releaseを`--c-api-version`で指定できるようになります。主な用途はこのvoicevox\_coreリポジトリでの内部利用です ([#1315])。
 
@@ -79,6 +79,8 @@
     - \[追加\]: `ndarray@0.17`
     - \[追加\]: `git+https://github.com/pykeio/ort.git?rev=94417081c47f47f5a7d6a92ce94bb38fda10019f#ort@2.0.0-rc.12`
     - \[変更\]: `indexmap@2`: `^2.6.0` → `^2.13.0`
+- \[C\] \[BREAKING\] `voicevox_synthesizer_load_voice_model`に引数`VoicevoxLoadVoiceModelOptions options`が追加されます ([#1337])。
+- \[Java\] \[BREAKING\] `Synthesizer#load_voice_model`がビルダースタイルになります ([#1337])。
 
 ### Security
 
@@ -1511,6 +1513,7 @@ Windows版ダウンローダーのビルドに失敗しています。
 [#1315]: https://github.com/VOICEVOX/voicevox_core/pull/1315
 [#1323]: https://github.com/VOICEVOX/voicevox_core/pull/1323
 [#1331]: https://github.com/VOICEVOX/voicevox_core/pull/1331
+[#1337]: https://github.com/VOICEVOX/voicevox_core/pull/1337
 [#1349]: https://github.com/VOICEVOX/voicevox_core/pull/1349
 [#1350]: https://github.com/VOICEVOX/voicevox_core/pull/1350